+++ examples/inverse/plot_lcmv_beamformer.py
@@ -58,9 +58,10 @@ forward = mne.read_forward_solution(fname_fwd)
 forward = mne.convert_forward_solution(forward, surf_ori=True)
 
 # Compute regularized noise and data covariances
-noise_cov = mne.compute_covariance(epochs, tmin=tmin, tmax=0, method='shrunk')
+noise_cov = mne.compute_covariance(epochs, tmin=tmin, tmax=0, method='shrunk',
+                                   rank=None)
 data_cov = mne.compute_covariance(epochs, tmin=0.04, tmax=0.15,
-                                  method='shrunk')
+                                  method='shrunk', rank=None)
 evoked.plot(time_unit='s')
 
 ###############################################################################